Put all* (* = enkelt.py, run_enkelt.py & EnkeltModules.py) into one file. Some smaller fixes. Added error message for when a import-module is not found
